requirements.txt
discord.py
emoji
apscheduler